+++ test/suites/PageTest.php
@@ -5,6 +5,7 @@
 
 namespace HeadlessChromium\Test;
 
+use HeadlessChromium\BrowserFactory;
 use HeadlessChromium\Communication\Connection;
 use HeadlessChromium\Communication\Session;
 use HeadlessChromium\Communication\Socket\MockSocket;
@@ -25,5 +26,22 @@ class PageTest extends BaseTestCase
         $page = new Page($target, []);
 
         $this->assertSame($session, $page->getSession());
+    }
+
+    public function testSetViewport()
+    {
+        $factory = new BrowserFactory();
+
+        $browser = $factory->createBrowser([
+            'windowSize' => [500, 500]
+        ]);
+
+        $page = $browser->createPage();
+
+        $page->setViewport(100, 300)->await();
+
+        $response = $page->evaluate('[window.innerWidth, window.innerHeight]')->getReturnValue();
+
+        $this->assertEquals([100, 300], $response);
     }
 }